requirements.txt
Flask==3.1.2
flask-cors==4.0.1
pymongo==4.6.1
python-dotenv==1.0.1
pytest==8.3.2
Werkzeug==3.1.3
temporalio==1.6.0